[bug-fix] ret is int
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@1165 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/teld/losmandy/losmandy.c
@@ -232,7 +232,7 @@ unlock:
 int
 tel_write_read (char *buf, int wcount, char *rbuf, int rcount)
 {
-  size_t ret;
+  int ret;
   ret = tel_write_read_no_reset (buf, wcount, rbuf, rcount);
   if (ret <= 0)
     {
